Add Stores Requests and Modules

+++ back-end/Routes/Menu.js
@@ -2,24 +2,48 @@
 import express from "express";
 import Menu from "../Schemas/Menu.js";
 import Category from "../Schemas/Category.js";
+import Store from "../Schemas/Store.js";
 
 // Variables
 const router = express.Router();
 router.use(express.json());
 
 // Create Menu
-router.post("/create-menu", async (req, res) => {
+router.post("/:storeId/create-menu", async (req, res) => {
   try {
     // Variables
+    const storeId = req.params.storeId;
+    const store = await Store.findById(storeId);
     const { name, doesDelivery, doesCollection } = req.body;
+
+    if (!store) {
+      return res.status(500).json({ message: "Store hasn't been found" });
+    }
+
+    // Check if the store already has a menu
+    if (store.menu) {
+      return res.status(400).json({
+        error: "This store already has a menu. You can't create another one.",
+      });
+    }
 
     // Create new menu
     const newMenu = new Menu({ doesDelivery, doesCollection, name });
 
-    // Save the menu into the database
-    const menu = await newMenu.save();
+    // Save the new category to get its _id
+    const createdMenu = await newMenu.save();
 
-    return res.status(201).json(menu);
+    // Assign the new menu's _id to the store's menu field
+    store.menu = createdMenu._id;
+
+    // Save the Menu inside the Store
+    await store.save();
+
+    return res.status(201).json({
+      message: `Menu has been created succesfully: `,
+      newMenu,
+      store,
+    });
   } catch (error) {
     console.error("Error creating menu:", error);
     return res.status(500).json({ error: "Unable to create the menu" });
@@ -27,13 +51,21 @@ router.post("/create-menu", async (req, res) => {
 });
 
 // Read menu
-router.get("/:menuId", async (req, res) => {
+router.get("/:storeId/:menuId", async (req, res) => {
   try {
     // Variables
     const menuId = req.params.menuId;
+    const storeId = req.params.storeId;
+
+    // Find the store by its ID
+    const store = await Store.findById(storeId);
 
     // Find the menu by its ID
     const menu = await Menu.findById(menuId);
+
+    if (!store) {
+      return res.status(404).json({ error: "Store not found." });
+    }
 
     if (!menu) {
       // If doesn't exist, send a 404 response
@@ -48,55 +80,64 @@ router.get("/:menuId", async (req, res) => {
   }
 });
 
-// Get all the menus
-router.get("/rq/get-menus", async (req, res) => {
-  try {
-    // Variables
-    const menus = await Menu.find({});
-
-    return res.json({ menus });
-  } catch (error) {
-    console.error("Error fetching data: ", error);
-    return res.status(500).json({ error: "Failed to retrieve data" });
-  }
-});
-
 // Update menu
-router.put("/:menuId", async (req, res) => {
+router.put("/:storeId/:menuId", async (req, res) => {
   try {
     // Variables
     const menuId = req.params.menuId;
+    const storeId = req.params.storeId;
+
+    const store = await Store.findById(storeId);
+    const menu = await Menu.findById(menuId);
+
+    if (!store) {
+      return res.status(404).json({ message: "Store not found" });
+    }
+
+    if (!menu) {
+      return res.status(404).json({ message: "Menu not found" });
+    }
 
     // Extract the updated menu data from the request body
-    const updatedMenuData = req.body;
+    const updatedMenuData = await req.body;
 
     // Attemp to find and update the menu by its ID
-    const menu = await Menu.findByIdAndUpdate(menuId, updatedMenuData, {
+    const menuUpdated = await Menu.findByIdAndUpdate(menuId, updatedMenuData, {
       new: true,
     });
 
-    if (menu) {
+    if (menuUpdated) {
       // If the menu was found and updated, respond with the updated menu
       return res
         .status(200)
-        .json({ message: "Menu updated successfully", menu: menu });
+        .json({ message: "Menu updated successfully", menu: menuUpdated });
     } else {
       return res.status(404).json({ error: "Menu not found" });
     }
   } catch (error) {
+    console.error("Error updating menu: ", error);
     // Handle any potential errors that might occur during the update
     return res.status(500).json({ error: "Internal server error" });
   }
 });
 
 // Delete Menu
-router.delete("/:menuId", async (req, res) => {
+router.delete("/:storeId/:menuId", async (req, res) => {
   try {
     // Variables
+    const storeId = req.params.storeId;
     const menuId = req.params.menuId;
+    const store = await Store.findById(storeId);
+
+    if (!store) {
+      return res.status(404).json({ message: "Store not found" });
+    }
+
     const menu = await Menu.findByIdAndDelete(menuId);
 
     if (menu) {
+      store.menu = undefined;
+      await store.save();
       const categories = menu.categories;
 
       if (categories.length > 0) {
@@ -109,6 +150,7 @@ router.delete("/:menuId", async (req, res) => {
         message:
           "Menu and its associated categories have been deleted successfully",
         menu: menu,
+        store: store,
       });
     }
 
@@ -121,7 +163,7 @@ router.delete("/:menuId", async (req, res) => {
 });
 
 // Read categories of the menu
-router.get("/:menuId/categories", async (req, res) => {
+router.get("/:storeId/:menuId/categories", async (req, res) => {
   try {
     // Variables
     const menuId = req.params.menuId;
